+++ CHANGELOG.md
@@ -11,10 +11,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - 外部決済サービス認証情報をプロダクト検索から取得するように調整
+- update @chevre/api-nodejs-client
+- update @motionpicture/gmo-service
+- update @pecorino/api-nodejs-client
 
 ### Deprecated
 
 ### Removed
+
+- 管理者としての口座開設サービスを削除
 
 ### Fixed